Add new installation routine
There is just username and SSH key in the installation for now.

Installation is interactive, non-destructive, but can
overwrite values if chosen. Steps can be skipped and whole
installation can be cancelled or aborted.
In case of abort, #already installed is written into
config so installation is not run again.

To restart installation this commented out info line should
be removed and broadexec restarted.